+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - Segments list: Use DpSearchField for custom search
 - DpInlineNotification: Set margin from outside the component (instead of inside)
 - Rework Tags-Lists: Now works with APi2.0 and is extendable by addons.
+- Enable the DELETE method for the AdministratableUser resource type
 - TagsList: trigger an update request only when the title is modified
 
 ## 2.20.0